Logout Dialog Fix: Vom Homescreen entfernt
