+++ client/src/components/DropdownMenu.tsx
@@ -15,6 +15,7 @@ import FolderIcon from "@mui/icons-material/Folder";
 import theme from "../theme/theme";
 import { useAuth } from "../contexts/AuthContext";
 import LoginReminderDialog from "./dialogs/LoginReminderDialog";
+import LogoutDialog from "../components/dialogs/Dialog";
 
 const MENU_ITEM_SX = {
   borderRadius: 999,
@@ -42,10 +43,11 @@ export default function DropMenu({
 }) {
   const navigate = useNavigate();
   const location = useLocation();
-  const { user, isLoading } = useAuth();
+  const { user, isLoading, logout } = useAuth();
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
   const open = Boolean(anchorEl);
   const [isDialogOpen, setIsDialogOpen] = useState(false);
+  const [isLogoutOpen, setIsLogoutOpen] = useState(false);
   const [intendedDestination, setIntendedDestination] = useState("");
 
   const handleMenuClick = (event: React.MouseEvent<HTMLElement>) => {
@@ -59,8 +61,7 @@ export default function DropMenu({
   const handleLoginLogout = () => {
     handleMenuClose();
     if (user) {
-      sessionStorage.setItem("showLogoutConfirmation", "true");
-      navigate("/");
+      setIsLogoutOpen(true);
     } else {
       navigate("/login");
     }
@@ -84,6 +85,22 @@ export default function DropMenu({
   const handleLoginClick = () => {
     setIsDialogOpen(false);
     navigate("/login", { state: { redirectTo: intendedDestination } });
+  };
+
+  const handleConfirmLogout = async () => {
+    setIsLogoutOpen(false);
+
+    try {
+      await logout();
+      sessionStorage.setItem("showLogoutConfirmation", "true");
+      navigate("/");
+    } catch (error) {
+      console.error("Logout failed:", error);
+    }
+  };
+
+  const handleCancelLogout = () => {
+    setIsLogoutOpen(false);
   };
 
   const menuItems: MenuItemConfig[] = [
@@ -170,6 +187,16 @@ export default function DropMenu({
         onClose={() => setIsDialogOpen(false)}
         onLoginClick={handleLoginClick}
         onProceedNavigation={handleProceedNavigation}
+      />
+
+      <LogoutDialog
+        open={isLogoutOpen}
+        onClose={() => handleCancelLogout()}
+        title="Möchtest du dich wirklich ausloggen?"
+        text=""
+        cancelLabel="NEIN"
+        confirmLabel="JA"
+        onConfirm={() => handleConfirmLogout()}
       />
     </>
   );
